refactor(cli): use getattr() for builtins access
Direct builtins.* access fails in some contexts; getattr() with
string lookup is more robust across import scenarios.

getattr(builtins, "howdy_user") and getattr(builtins, "howdy_args")
avoid AttributeError during CLI module imports when howdy_user
is not set. Matches pattern already used elsewhere in the codebase.

+++ howdy/src/cli/test.py
@@ -286,8 +286,8 @@ try:
             # are captured and even after a delay it does not
             # always work. Setting exposure at every frame is
             # reliable though.
-            video_capture.internal.set(cv2.CAP_PROP_AUTO_EXPOSURE, 1.0)  # 1 = Manual
-            video_capture.internal.set(cv2.CAP_PROP_EXPOSURE, float(exposure))
+            video_capture.internal.set(cv2.CAP_PROP_AUTO_EXPOSURE, 1)  # 1 = Manual
+            video_capture.internal.set(cv2.CAP_PROP_EXPOSURE, int(exposure))
 
 except KeyboardInterrupt:
     cv2.destroyAllWindows()
